# src/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UnisonHT: Raspberry Pi IR Hat",
        "version": "1.0.0",
        "description": "UnisonHT web service for Raspberry Pi IR Hat.",
        "license": {
            "name": "MIT",
            "url": "https://raw.githubusercontent.com/joeferner/unisonht-raspberry-pi-ir-hat/master/LICENSE.txt"
        }
    },
    "paths": {
        "/api/v1/config": {
            "get": {
                "operationId": "getConfig",
                "summary": "Get configuration",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "couple remotes and buttons": {
                                        "value": {
                                            "remotes": {
                                                "test1": {
                                                    "buttons": {
                                                        "volumeUp": {
                                                            "signal": "100,200,300"
                                                        },
                                                        "volumeDown": {
                                                            "signal": "100,200,400"
                                                        }
                                                    }
                                                },
                                                "test2": {
                                                    "buttons": {
                                                        "channelUp": {
                                                            "signal": "100,200,300"
                                                        },
                                                        "channelDown": {
                                                            "signal": "100,200,400"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/transmit/{remoteName}/{buttonName}": {
            "post": {
                "operationId": "transmit",
                "summary": "Transmit a remote + button IR signal",
                "parameters": [
                    {
                        "name": "remoteName",
                        "in": "path",
                        "description": "Name of the remote found in config",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "buttonName",
                        "in": "path",
                        "description": "Name of the button found on the remote in config",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "ok": {
                                        "value": {}
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Either the remote or button was not found"
                    },
                    "408": {
                        "description": "Timeout communicating with device"
                    }
                }
            }
        }
    }
}